Added card flip feature.
Essentially, how it works is that each card has an associated "Other Side". In UnoDeck, we generate two decks: one light, one dark. We then shuffle them, and link every light card to a dark card. UnoDeck, Player, and View all implement an additional interface called "flip". When the flip method is called in Game, all the flip listeners will handle it in their own way. UnoDeck will swap which deck is active, the light or the dark, and player will "turn over" all the cards in their hand, substituting each card for the associated "other side" card. A little message appears, and the illusion is complete.